requirements.txt
streamlit==1.28.0
pypdf==4.0.1
python-pptx==0.6.21
groq==0.4.1
python-dotenv==1.0.0
reportlab==4.0.4
pdfplumber==0.10.0 